requirements.txt
pymongo
paho-mqtt
matplotlib
python-dotenv 